+++ config/prebuild.js
@@ -1,9 +1,11 @@
-var fs = require("fs");
-var path = require("path")
-var deleteFolderRecursive = function (path) {
+const fs = require("fs");
+const pathLib = require("path");
+const Utility = require("./util/Utility");
+
+const deleteFolderRecursive = function (path) {
     if (fs.existsSync(path)) {
         fs.readdirSync(path).forEach(function (file, index) {
-            var curPath = path + "/" + file;
+            var curPath = pathLib.normalize(path + "/" + file);
             console.log("delete: ", curPath);
             if (fs.lstatSync(curPath).isDirectory()) { // recurse
                 deleteFolderRecursive(curPath);
@@ -16,5 +18,5 @@ var deleteFolderRecursive = function (path) {
 };
 
 process.argv.slice(2).forEach(function (val, index, array) {
-    deleteFolderRecursive(path.resolve(val));
+    deleteFolderRecursive(pathLib.join(Utility.projectDir, val));
 });